test vecsort

+++ assignment_2/mergesort/sequential/merge.c
@@ -15,15 +15,15 @@ int debug = 0;
 void top_down_mergesort(int *b, long l, int *a){
     if(l<=1)
         return;
-    int num_rhs = l/2;
-    int num_lhs = l-num_rhs;
+    long num_rhs = l/2;
+    long num_lhs = l-num_rhs;
     int *mid_a = a+num_lhs;
     int *mid_b = b+num_lhs;
     top_down_mergesort(a, num_lhs, b);
     top_down_mergesort(mid_a, num_rhs, mid_b);
-    int i=0,j=0;
+    long i=0,j=0;
 
-    for(int k = 0; k < l; k++) {
+    for(long k = 0; k < l; k++) {
         if (i < num_lhs && (j >= num_rhs || b[i] <= mid_b[j])) {
             a[k] = b[i];
             i++;
